Code cleanup, improve editor functionality.
Reverted back to using the OnValidate method for Forgelight Integration objects. This now saves Light and Object info when prefabs are created, or the scene is saved.
Added headers to ZoneLight object to make each section more noticeable in the editor.
Refactored "ConvertTransform" method to "InvertTransform" to better match it's functionality.
Fixed spotlights not being rotated to the correct direction.

KNOWN BUGS:
MAJOR: Exporting a zone file corrupts the level if tried to be re-imported into the editor. It will work in-game, but attempting to re-import the level will result in object positions being incorrect.
    -    This seems to be an issue with the transform translation, and should be fixed next commit.