Fix Entity tool not parenting, non-scene objects affected by object finds, object rotation showing incorrectly in editor, remove test scene.
Fixed an issue where the entity tool would not parent any objects or lights if there was none of the other type.
Fixed an issue where prefabs, and other gameobjects not in the scene, would be affected by import/export/culling.
Fixed the rotations of zone objects would import and export correctly, but appear distorted in-editor.
Fixed the label on the progress bar which would show no Zone name when loading zones.